Beginning refactoring | Changed sendToBackend to not allow multiple simulation attempts while a simulation is running
